+++ app/src/main/java/android/example/blowonboredom/ui/MainActivity.kt
@@ -1,12 +1,9 @@
-package android.example.blowonboredom
+package android.example.blowonboredom.ui
 
 
-import android.example.blowonboredom.fragments.RandomActivity
+import android.example.blowonboredom.R
 import android.os.Bundle
-import android.view.View
-import android.view.animation.LinearInterpolator
 
-import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.navigation.NavController
 import androidx.navigation.fragment.NavHostFragment
@@ -14,11 +11,7 @@ import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.NavigationUI
 import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
-import androidx.recyclerview.widget.DefaultItemAnimator
-import androidx.recyclerview.widget.DiffUtil
 import com.google.android.material.bottomnavigation.BottomNavigationView
-import com.yuyakaido.android.cardstackview.*
-import kotlin.random.Random
 
 
 class MainActivity : AppCompatActivity() {
