+++ app/src/main/java/android/example/blowonboredom/ui/fragments/DetailActivityFragment.kt
@@ -1,10 +1,8 @@
-package android.example.blowonboredom.fragments
+package android.example.blowonboredom.ui.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.example.blowonboredom.R
 import androidx.core.view.ViewCompat
