+++ app/src/main/res/layout/fragment_detail_activity.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.DetailActivityFragment">
+    tools:context=".ui.fragments.DetailActivityFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
