+++ app/src/main/res/layout/fragment_favorite_activities.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.FavoriteActivitiesFragment">
+    tools:context=".ui.fragments.FavoriteActivitiesFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
