+++ app/src/main/res/layout/fragment_random_activities.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".fragments.RandomActivitiesFragment">
+    tools:context=".ui.fragments.RandomActivitiesFragment">
     <com.yuyakaido.android.cardstackview.CardStackView
         android:id="@+id/card_stack_view"
         android:layout_width="0dp"
