+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,12 +7,12 @@
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="android.example.blowonboredom.fragments.HomeFragment"
+        android:name="android.example.blowonboredom.ui.fragments.HomeFragment"
         android:label="fragment_home"
         tools:layout="@layout/fragment_home" />
     <fragment
         android:id="@+id/favoriteActivitiesFragment"
-        android:name="android.example.blowonboredom.fragments.FavoriteActivitiesFragment"
+        android:name="android.example.blowonboredom.ui.fragments.FavoriteActivitiesFragment"
         android:label="fragment_favorite_activities"
         tools:layout="@layout/fragment_favorite_activities" >
 
@@ -26,7 +26,7 @@
     </fragment>
     <fragment
         android:id="@+id/randomActivitiesFragment"
-        android:name="android.example.blowonboredom.fragments.RandomActivitiesFragment"
+        android:name="android.example.blowonboredom.ui.fragments.RandomActivitiesFragment"
         android:label="fragment_random_activities"
         tools:layout="@layout/fragment_random_activities" >
 
@@ -36,7 +36,7 @@
     </fragment>
     <fragment
         android:id="@+id/detailActivityFragment"
-        android:name="android.example.blowonboredom.fragments.DetailActivityFragment"
+        android:name="android.example.blowonboredom.ui.fragments.DetailActivityFragment"
         android:label="fragment_detail_activity"
         tools:layout="@layout/fragment_detail_activity" >
         <action
